Refactor(SE-62): Adjustments to folder structures
