+++ app/dashboard/rates/_components/RatesTableSummary.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { usePeakSeasonRate } from "@/hooks/usePeakSeasonRate";
+import { usePeakSeasonRate } from "@/hooks/reports/usePeakSeasonRate";
 import RateTable from "@/app/dashboard/rates/_components/RateTable";
 
 const RatesTableSummary = () => {
