+++ hooks/reports/usePeakSeasonRate.ts
