+++ hooks/reports/useRoomAvailability.ts
