+++ hooks/reports/useRoomAvailabilityCalendarConfig.tsx
